refactor(router): 更新路由组件路径
- 修改 writecode、aiprocess 和 problemdisplay 路由的组件路径
- 删除 auth 目录下的相关组件文件
- 调整 views 目录结构，将问题相关组件移至 problem 目录
