+++ src/types/AiProcess.ts
@@ -1,16 +1,3 @@
-
-// //
-// export interface ProblemInfo {
-//    title:string;
-//    description:string;
-//    input:string;
-//    output:string;
-//    sample_input:string;
-//    sampel_output:string;
-//    hint:string;
-//    tags:string[];
-//    solution:string;
-// }
 export interface CodeSubmit {
     language_id:number;
     sourse_code:string;
